changed int -> long

+++ server/hitchbot-secure-api/hitchbot-secure-api/Controllers/ReturnObjects/GenericHitchBot.cs
@@ -11,7 +11,7 @@ namespace hitchbot_secure_api.Controllers.ReturnObjects
     {
         public int HitchBotId { get; set; }
 
-        public int? TimeUnix { get; set; }
+        public long? TimeUnix { get; set; }
 
         public string Time { get; set; }
